+++ camera.py
@@ -1,0 +1,7 @@
+import camera.takeImage
+import picamera
+
+
+camera.takeImage.test()
+
+with picamera.PiCamera() as camera:
